update to support musa

+++ projects/mmdet3d_plugin/models/detectors/bevdet.py
@@ -71,7 +71,8 @@ class BEVDet(CenterPoint):
         keyego2global = ego2globals[:, 0,  ...].unsqueeze(1)    # (B, 1, 4, 4)
         global2keyego = torch.inverse(keyego2global.double())   # (B, 1, 4, 4)
         sensor2keyegos = \
-            global2keyego @ ego2globals.double() @ sensor2egos.double()     # (B, N_views, 4, 4)
+             global2keyego.float() @ ego2globals.float() @ sensor2egos.float() 
+#            global2keyego @ ego2globals.double() @ sensor2egos.double()     # (B, N_views, 4, 4)
         sensor2keyegos = sensor2keyegos.float()
 
         return [imgs, sensor2keyegos, ego2globals, intrins,
